Fixed issue: Array(numbers) does not allow to set decimal step width on certain browsers

+++ application/helpers/questionHelper.php
@@ -649,7 +649,7 @@ class questionHelper
             "types"=>":",
             'category'=>gT('Display'),
             'sortorder'=>111,
-            'inputtype'=>'integer',
+            'inputtype'=>'float',
             'default'=>'',
             "help"=>gT('Step value'),
             "caption"=>gT('Step value')
